Overhauld the generate emails function
The generate emails function now operates off of email verification.
As of now, it will evaluate the results the email bot returns and
determine whether the email format is valid or not. If it deems
one of the email formats valid it will apply that format to the contact
and log the format in a json file. However, if the all the email formats
are deemed invalid, then the format for that company will be logged as
invalid in the json file. The json file is used for other contacts
to save api requests and searches on the email bot. If a company
already has a valid format, then it doesnt need to re-search and
validate it. The opposit is also true for invalid formats.

As of now the vaildation only works off of the chance specified in
the search results. i.e. "The email format janedoe@rocketreach.com
is used 87% of the time". Future implimentations will hopefully use
an email verifier api.